+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/package-info.java
@@ -21,7 +21,7 @@
  * <ul>
  * <li>{@link is.codion.swing.framework.ui.component.EntitySearchField#SEARCH_INDICATOR}
  * <li>{@link is.codion.swing.framework.ui.component.EntitySearchField#SEARCH_PROGRESS_BAR_DELAY}
- * <li>{@link is.codion.swing.framework.ui.component.DefaultEditComponentFactory#DEFAULT_TEXT_FIELD_COLUMNS}
+ * <li>{@link is.codion.swing.framework.ui.component.DefaultEditComponent#DEFAULT_TEXT_FIELD_COLUMNS}
  * </ul>
  */
 @org.jspecify.annotations.NullMarked